requirements.txt
pandas
psycopg2-binary
PyYAML
python-dotenv
pytest
